app/src/main/java/org/xwiki/android/sync/utils/StringUtils.java
/*
 * See the NOTICE file distributed with this work for additional
 * information regarding copyright ownership.
 *
 * This is free software; you can redistribute it and/or modify it
 * under the terms of the GNU Lesser General Public License as
 * published by the Free Software Foundation; either version 2.1 of
 * the License, or (at your option) any later version.
 *
 * This software is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this software; if not, write to the Free
 * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
 * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 */
package org.xwiki.android.sync.utils;

import java.text.ParseException;
import java.text.SimpleDateFormat;
import java.util.Date;
import java.util.TimeZone;
import java.util.regex.Matcher;
import java.util.regex.Pattern;

/**
 * String Time Utils
 */
public class StringUtils {
    private final static Pattern emailer = Pattern
            .compile("\\w+([-+.]\\w+)*@\\w+([-.]\\w+)*\\.\\w+([-.]\\w+)*");

    /**
     * isEmpty
     * whether input is '', null,'\t' '\r' '\n'  or not.
     * '', null, return ture
     * many '\t' '\r' '\n' return true;
     * other false
     */
    public static boolean isEmpty(CharSequence input) {
        if (input == null || "".equals(input))
            return true;
        for (int i = 0; i < input.length(); i++) {
            char c = input.charAt(i);
            if (c != ' ' && c != '\t' && c != '\r' && c != '\n') {
                return false;
            }
        }
        return true;
    }

    /**
     * whether email is valid or not
     */
    public static boolean isEmail(CharSequence email) {
        if (isEmpty(email))
            return false;
        return emailer.matcher(email).matches();
    }

    /**
     * return system time with format
     */
    public static String getDataTime(String format) {
        SimpleDateFormat df = new SimpleDateFormat(format);
        return df.format(new Date());
    }

    /**
     * isIpAddress
     * whether a charSequence is a valid ip address.
     * @param addr
     * @return
     */
    public static boolean isIpAddress(CharSequence addr) {
        if (addr == null) return false;
        if (addr.length() < 7 || addr.length() > 15 || "".equals(addr)) {
            return false;
        }
        String rexp = "([1-9]|[1-9]\\d|1\\d{2}|2[0-4]\\d|25[0-5])(\\.(\\d|[1-9]\\d|1\\d{2}|2[0-4]\\d|25[0-5])){3}";
        Pattern pat = Pattern.compile(rexp);
        Matcher mat = pat.matcher(addr);
        return mat.find();
    }

    /**
     * isDomainAddress
     * whether a charSequence is a valid domain address
     *
     * @param addr
     * @return
     */
    public static boolean isDomainAddress(CharSequence addr) {
        if (addr == null || addr.length() < 3) return false;
        String rexp = "^(?=^.{3,255}$)[a-zA-Z0-9][-a-zA-Z0-9]{0,62}(\\.[a-zA-Z0-9][-a-zA-Z0-9]{0,62})+$";
        Pattern pat = Pattern.compile(rexp);
        Matcher mat = pat.matcher(addr);
        return mat.find();
    }

    /**
     * iso8601ToDate
     * Iso8601 String To Date
     * 2011-09-24T19:45:31+02:00
     */
    public static Date iso8601ToDate(String iso8601) {
        if(isEmpty(iso8601)) return null;
        //TODO it's a problem
        // in android need "yyyy-MM-dd'T'HH:mm:ssZ"
        // in junit4 need  "yyyy-MM-dd'T'HH:mm:ssX" iso8601
        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ");
        sdf.setTimeZone(TimeZone.getTimeZone("GMT"));
        try {
            return sdf.parse(iso8601);
        } catch (ParseException e) {
            e.printStackTrace();
        }
        return null;
    }

    /**
     * dateToIso8601String
     *   Date to Iso8601 String
     *   2011-09-24T19:45:31+02:00
     * @param date
     * @return String
     */
    public static String dateToIso8601String(Date date) {
        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ssZ");
        sdf.setTimeZone(TimeZone.getTimeZone("GMT"));
        return sdf.format(date);
    }

}
